% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pregnancy_start_end_hierarchy.R
\name{apply_pregnancy_start_end_hierarchy}
\alias{apply_pregnancy_start_end_hierarchy}
\title{Apply Pregnancy Start and End Hierarchy to Patients}
\usage{
apply_pregnancy_start_end_hierarchy(outcome_hierarchy_final)
}
\arguments{
\item{outcome_hierarchy_final}{A data frame containing patient data with variables such as enrolid, svcdate, code, outcome, trimester_code_clean, and others.}
}
\value{
A data frame with the applied pregnancy start and end hierarchy logic, containing variables such as enrolid, svcdate, code, outcome, trimester_code_clean, and others.
}
\description{
This function estimates the pregnancy start and end dates for Commercial Claims or Medicaid MarketScan patients and applies hierarchy logic to resolve conflicts.
It processes the data through multiple stages, resolving conflicts for ectopic, abortion, stillbirth, preterm, delivery, unclassified loss, postpartum, and trimester outcomes.
}
\examples{
\dontrun{
trimester_hierarchy_final <- apply_trimester_hierarchy(patients)
outcome_hierarchy_final <- apply_outcome_hierarchy(trimester_hierarchy_final)
pregnancy_data_cleaned <- apply_pregnancy_start_end_hierarchy(outcome_hierarchy_final)
}
}
