% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_trimester_hierarchy.R
\name{apply_trimester_hierarchy}
\alias{apply_trimester_hierarchy}
\title{Apply Trimester Hierarchy to Patients}
\usage{
apply_trimester_hierarchy(patients)
}
\arguments{
\item{patients}{A data frame containing patient data with the format of the exported dataset extraction example dataset.}
}
\value{
A data frame with the applied trimester hierarchy logic, containing variables such as enrolid, svcdate, code, claims_file, description, final_trimester_code, secondary_outcome, srv_year, outcome, pdx, and trimester_code_clean.
}
\description{
Identify duplicate trimester codes for the same service day and apply hierarchy logic to resolve conflicts.
Group procedures performed on eligible women by enrolid and date of procedure.
Multiple conflicting trimester codes could occur on the same day, so the trimester group code for that service day needs to be reassigned based on the logic within the function.
}
\examples{
\dontrun{
patients <- arrow::read_parquet("patients.parquet")
trimester_hierarchy_final <- apply_trimester_hierarchy(patients)
}
}
