% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_age_variables}
\alias{calculate_age_variables}
\title{Calculate additional age variables}
\usage{
calculate_age_variables(.data)
}
\arguments{
\item{.data}{Combined ICD9/ICD10, PROC, and CPT as a single dataset}
}
\value{
Modified dataset
}
\description{
Calculate additional age variables
}
