% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_NAs_in_trimester_code_clean}
\alias{drop_NAs_in_trimester_code_clean}
\title{Drop NAs helper}
\usage{
drop_NAs_in_trimester_code_clean(trimester_data)
}
\arguments{
\item{trimester_data}{trimester_data}
}
\value{
filtered trimester_data
}
\description{
Drop NAs helper
}
