% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extraction_example}
\alias{extraction_example}
\title{Extraction Example}
\format{
A data frame with 14 rows and 11 columns:
\describe{
\item{enrolid}{Enrolment ID as a double}
\item{dobyr}{Year of birth as a double}
\item{code}{Medical code as a character string}
\item{description}{Description of the medical code}
\item{pdx}{Primary diagnosis indicator as a double}
\item{claims_file}{Claims file type as a character string}
\item{svcdate}{Service date as a character string}
\item{final_trimester_code}{Numeric code representing the final trimester as a double}
\item{outcome}{Primary outcome related to the medical code}
\item{secondary_outcome}{Secondary outcome related to the medical code as a logical}
\item{srv_year}{Service year as a double}
}
}
\usage{
extraction_example
}
\description{
Example dataset showing what the results of the extraction step should look like.
}
\keyword{datasets}
