% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_eligible_women}
\alias{filter_eligible_women}
\title{ID women that fit inclusion criteria of age and time period}
\usage{
filter_eligible_women(
  .data,
  .variables_to_keep,
  .age_start,
  .age_end,
  .last_yr,
  .presvcdate,
  .svcend,
  .claims_file
)
}
\arguments{
\item{.data}{MarketScan dataset}

\item{.variables_to_keep}{Variable names that should appear in the returned dataset}

\item{.age_start}{int, lower age boundary for inclusion}

\item{.age_end}{int, upper age boundary for inclusion}

\item{.last_yr}{int, study end year. Last year of full data on MarketScan dataset}

\item{.presvcdate}{POSIXct, start of study period minus a certain number of days to capture pregnancy start dates}

\item{.svcend}{POSIXct, end of study period}

\item{.claims_file}{chr, "I" for inpatient or "O" for outpatient}
}
\value{
Subset dataset
}
\description{
ID women that fit inclusion criteria of age and time period
}
