% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_matches_in_ICD}
\alias{find_matches_in_ICD}
\title{Go through the various "diagnosis" fields in the dataset, looking for matches with the ICD codes}
\usage{
find_matches_in_ICD(
  .data,
  inpat,
  .dxver,
  .icd_data,
  .icd_value,
  .variables_to_keep
)
}
\arguments{
\item{.data}{MarketScan dataset}

\item{inpat}{logical, for inpat or outpat claims}

\item{.dxver}{procedure code version}

\item{.icd_data}{dataset with icd codes}

\item{.icd_value}{icd value to match}

\item{.variables_to_keep}{Variable names that should appear in the returned dataset}
}
\value{
Dataset
}
\description{
Go through the various "diagnosis" fields in the dataset, looking for matches with the ICD codes
}
