% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_diagnosis_pregnancy_codes}
\alias{icd10_diagnosis_pregnancy_codes}
\title{ICD-10 Diagnosis Pregnancy Codes}
\format{
A data frame with 3,132 rows and 6 columns:
\describe{
\item{icd10cm}{ICD-10-CM code as a character string}
\item{icd10cm_ihd}{ICD-10-CM code with a period as a character string}
\item{description}{Description of the ICD-10-CM code}
\item{outcome}{Primary pregnancy outcome related to the ICD-10-CM code}
\item{secondary_outcome}{Secondary pregnancy outcome related to the ICD-10-CM code}
\item{final_trimester_code}{Numeric code representing the final trimester}
}
}
\source{
Created internally within GSK
}
\usage{
icd10_diagnosis_pregnancy_codes
}
\description{
Pregnancy code list for ICD10
}
\keyword{datasets}
