% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_procedure_pregnancy_codes}
\alias{icd10_procedure_pregnancy_codes}
\title{ICD-10 Procedure Pregnancy Codes}
\format{
A data frame with 416 rows and 4 columns:
\describe{
\item{icd10_proc}{ICD-10 Procedure code as a character string}
\item{description}{Description of the ICD-10 Procedure code}
\item{outcome}{Primary pregnancy outcome related to the ICD-10 Procedure code}
\item{final_trimester_code}{Numeric code representing the final trimester}
}
}
\source{
Created internally within GSK
}
\usage{
icd10_procedure_pregnancy_codes
}
\description{
Pregnancy code list for ICD10 Procedure
}
\keyword{datasets}
