% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9_diagnosis_pregnancy_codes}
\alias{icd9_diagnosis_pregnancy_codes}
\title{ICD-9 Diagnosis Pregnancy Codes}
\format{
A data frame with 1,927 rows and 6 columns:
\describe{
\item{icd9cm}{ICD-9-CM code as a character string}
\item{icd9cm_ihd}{ICD-9-CM code with a period as a character string}
\item{description}{Description of the ICD-9-CM code}
\item{outcome}{Primary outcome related to the ICD-9-CM code}
\item{secondary_outcome}{Secondary outcome related to the ICD-9-CM code}
\item{final_trimester_code}{Numeric code representing the final trimester}
}
}
\source{
Created internally within GSK
}
\usage{
icd9_diagnosis_pregnancy_codes
}
\description{
Pregnancy code list for ICD9
}
\keyword{datasets}
