% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9_procedure_pregnancy_codes}
\alias{icd9_procedure_pregnancy_codes}
\title{ICD-9 Procedure Pregnancy Codes}
\format{
A data frame with 89 rows and 5 columns:
\describe{
\item{icd9_proc}{ICD-9 Procedure code as a character string}
\item{icd9_proc_ihd}{ICD-9 Procedure code with a period as a character string}
\item{description}{Description of the ICD-9 Procedure code}
\item{outcome}{Primary pregnancy outcome related to the ICD-9 Procedure code}
\item{final_trimester_code}{Numeric code representing the final trimester}
}
}
\source{
Created internally within GSK
}
\usage{
icd9_procedure_pregnancy_codes
}
\description{
Pregnancy code list for ICD9 Procedure
}
\keyword{datasets}
