% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_only_NAs_in_trimester_code_clean}
\alias{keep_only_NAs_in_trimester_code_clean}
\title{Keep NAs helper}
\usage{
keep_only_NAs_in_trimester_code_clean(trimester_data)
}
\arguments{
\item{trimester_data}{trimester_data}
}
\value{
filtered trimester_data
}
\description{
Keep NAs helper
}
