% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labor_delivery_codes}
\alias{labor_delivery_codes}
\title{Labour and Delivery Codes}
\format{
A data frame with 1,014 rows and 3 columns:
\describe{
\item{code}{Code as a character string}
\item{description}{Description of the code}
\item{file}{Code type as a character string (i.e. "ICD10DX","ICD10PC","ICD9DX" ,"ICD9PC", and "CPT")}
}
}
\source{
Created internally within GSK
}
\usage{
labor_delivery_codes
}
\description{
Pregnancy code list for Labor and Delivery
}
\keyword{datasets}
