% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_on_pregnancy_cpt_code}
\alias{match_on_pregnancy_cpt_code}
\title{Keep and return records of eligible women that match CPT pregnancy codes with additional info}
\usage{
match_on_pregnancy_cpt_code(
  .data,
  .cpt_data,
  .by_var_claims = "proc1",
  .by_var_cpt_code = "cpt_char",
  .variables_to_keep
)
}
\arguments{
\item{.data}{MarketScan dataset}

\item{.cpt_data}{CPT procedure codes dataset}

\item{.by_var_claims}{Variable in claims data to join by}

\item{.by_var_cpt_code}{Variable in CPT procedure codes dataset to join by}

\item{.variables_to_keep}{Variable names that should appear in the returned dataset}
}
\value{
Modified dataset
}
\description{
Keep and return records of eligible women that match CPT pregnancy codes with additional info
}
