% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_on_pregnancy_procedure_code}
\alias{match_on_pregnancy_procedure_code}
\title{Keep and return records of eligible women that match pregnancy procedure codes with additional info}
\usage{
match_on_pregnancy_procedure_code(
  .data,
  .proc_data,
  .by_var_claims,
  .by_var_proc_code,
  .variables_to_keep
)
}
\arguments{
\item{.data}{MarketScan dataset}

\item{.proc_data}{Procedure code dataset}

\item{.by_var_claims}{Variable in claims data to join by}

\item{.by_var_proc_code}{Variable in procedure code dataset to join by}

\item{.variables_to_keep}{Variable names that should appear in the returned dataset}
}
\value{
Modified dataset
}
\description{
Keep and return records of eligible women that match pregnancy procedure codes with additional info
}
