% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Nsundays}
\alias{Nsundays}
\title{Calculate Nsundays (Number of Sundays per week)}
\usage{
Nsundays(a, b)
}
\arguments{
\item{date1, date2}{Dates}
}
\value{
Number of Sundays between dates
}
\description{
Calculate Nsundays (Number of Sundays per week)
}
