% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregfindr-package.R
\docType{package}
\name{pregfindr-package}
\alias{pregfindr}
\alias{pregfindr-package}
\title{pregfindr: Standardised Pregnancy Identification And Trimester Distribution In MarketScan}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Identify pregnancies in a specific cut of the MarketScan dataset using the algorithm described in Sumner, K. M., Ehlinger, A., Georgiou, M. E., & Wurst, K. E. (2021). Development and evaluation of standardized pregnancy identification and trimester distribution algorithms in U.S. IBM MarketScan® Commercial and Medicaid data. Birth defects research, 113(19), 1357--1367. \url{https://doi.org/10.1002/bdr2.1954}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gsk-biostatistics.github.io/pregfindr/}
}

}
\author{
\strong{Maintainer}: Federico Concas \email{federico.x.concas@gsk.com} (\href{https://orcid.org/0000-0002-6434-6065}{ORCID})

Authors:
\itemize{
  \item Jaspreet Multani \email{jaspreet.k.multani@gsk.com}
}

Other contributors:
\itemize{
  \item GlaxoSmithKline Research & Development Limited [copyright holder, funder]
}

}
\keyword{internal}
