% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_parquet}
\alias{save_parquet}
\title{Save an R object as a Parquet file}
\usage{
save_parquet(obj, path = ".", save = FALSE)
}
\arguments{
\item{obj}{The R object to be saved as a Parquet file. This is typically a data frame.}

\item{path}{The directory where the Parquet file will be saved. Defaults to the current working directory.}

\item{save}{logical, if FALSE object is returned without saving}
}
\value{
NULL. The function prints a message indicating the file has been saved.
}
\description{
This function saves an R object (e.g., a data frame) as a Parquet file using the arrow package.
The file is saved in the specified directory with the object's name as the file name.
}
\examples{
my_dataframe <- data.frame(a = 1:5, b = letters[1:5])
save_parquet(my_dataframe)
}
\keyword{internal}
