% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{success_confirmation}
\alias{success_confirmation}
\title{Output a success confirmation message}
\usage{
success_confirmation(data = NULL)
}
\arguments{
\item{data}{An optional data frame (not used)}
}
\value{
No return value; the function outputs the success confirmation message
}
\description{
This function outputs a success confirmation message. It is intended to be used after a step in a process has been completed successfully.
}
