% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validation_error}
\alias{validation_error}
\title{Generate a validation error message}
\usage{
validation_error(list_of_errors, data = NULL, ...)
}
\arguments{
\item{list_of_errors}{A list of errors to include in the message}

\item{data}{An optional data frame that may have attached errors}

\item{...}{Additional arguments (not used)}
}
\value{
No return value; the function outputs the error message
}
\description{
This function generates a validation error message based on a list of errors.
If a data frame is provided and it has attached errors, these errors will be included in the message.
The message will include the total number of errors and the details of each error.
}
